… csrc/rocm/attention.cu