# Description
This is a PR to merge https://github.com/ROCm/vllm/tree/shsanyal_devpa_308_opt optimized `attention.cu` kernel into `llama_fp8_12062024` branch.

# CAVEAT
Currently the `attention.cu` kernel does not support `block size` of `32` and `head size` of `64`.
The vLLM model unittests are failing as it uses small models e.g. Gemma, Llama which has `head size` of `64`.

# Performance
The following is a `benchmark_throughput` results of `Llama-3.1-70B` with `fp8` dynamic quantization and `kv-cache-dtype` of `fp8_e4m3`.

Branch of vll-rocmfork | Dataset | Input length | Output length | Prompt numbers | Req/s | Total Tokens/s | Output Tokens/s | Req/s over main (%) | Total Tokens/s over main (%) | Output Tokens/s over main (%) | Req/s over llama_fp8 (%) | Total tokens/s over llama_fp8 (%) | OutputTokens/s over llama_fp8 (%) | Req/s over pagedattn-fp8 (%) | Total tokens/s over pagedattn-fp8 (%) | OutputTokens/s over  pagedattn-fp8 (%)
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
main | Random Sequence | 128 | 128 | 200 | 13.41 | 3432.02 | 1716.01 |   |   |   | 15.50387597 | 15.45787595 | 15.45748754 | 6.007905138 | 5.995243831 | 5.995243831
main | Random Sequence | 128 | 2048 | 200 | 0.68 | 1475.11 | 1388.33 |   |   |   | 1.492537313 | 1.522377993 | 1.521732772 | -38.73873874 | -38.93679291 | -38.93719679
main | Random Sequence | 2048 | 128 | 200 | 2.04 | 4436.77 | 260.99 |   |   |   | 61.9047619 | 62.34302734 | 62.34759891 | 48.90510949 | 48.43709748 | 48.44158799
main | Random Sequence | 2048 | 2048 | 200 | 0.29 | 1196.2 | 598.1 |   |   |   | 3.571428571 | 3.795359492 | 3.795359492 | -38.29787234 | -38.10859195 | -38.10859195
main | ShareGPT_V3_unfiltered_cleaned_split |   |   | 1000 | 5.98 | 2474.49 | 1186.82 |   |   |   | 9.926470588 | 10.06979196 | 10.07011426 | -9.531013616 | -9.527361274 | -9.527366977
llama-fp8-12062024 | Random Sequence | 128 | 128 | 200 | 11.61 | 2972.53 | 1486.27 | -13.42281879 | -13.38832524 | -13.38803387 |   |   |   | -8.221343874 | -8.195744155 | -8.195435313
llama-fp8-12062024 | Random Sequence | 128 | 2048 | 200 | 0.67 | 1452.99 | 1367.52 | -1.470588235 | -1.499549186 | -1.498923167 |   |   |   | -39.63963964 | -39.85246573 | -39.8524813
llama-fp8-12062024 | Random Sequence | 2048 | 128 | 200 | 1.26 | 2732.96 | 160.76 | -38.23529412 | -38.40203572 | -38.40377026 |   |   |   | -8.02919708 | -8.565769708 | -8.565578432
llama-fp8-12062024 | Random Sequence | 2048 | 2048 | 200 | 0.28 | 1152.46 | 576.23 | -3.448275862 | -3.656579167 | -3.656579167 |   |   |   | -40.42553191 | -40.37170028 | -40.37170028
llama-fp8-12062024 | ShareGPT_V3_unfiltered_cleaned_split |   |   | 1000 | 5.44 | 2248.11 | 1078.24 | -9.030100334 | -9.148551823 | -9.148817849 |   |   |   | -17.70045386 | -17.80429751 | -17.80454338
pagged-attn-fp8 | Random Sequence | 128 | 128 | 200 | 12.65 | 3237.9 | 1618.95 | -5.667412379 | -5.656144195 | -5.656144195 | 8.957795004 | 8.927412003 | 8.927045557 |   |   |  
pagged-attn-fp8 | Random Sequence | 128 | 2048 | 200 | 1.11 | 2415.71 | 2273.61 | 63.23529412 | 63.76473619 | 63.76581937 | 65.67164179 | 66.25785449 | 66.25789751 |   |   |  
pagged-attn-fp8 | Random Sequence | 2048 | 128 | 200 | 1.37 | 2988.99 | 175.82 | -32.84313725 | -32.63139626 | -32.63343423 | 8.73015873 | 9.368230783 | 9.368001991 |   |   |  
pagged-attn-fp8 | Random Sequence | 2048 | 2048 | 200 | 0.47 | 1932.74 | 966.37 | 62.06896552 | 61.5733155 | 61.5733155 | 67.85714286 | 67.70560367 | 67.70560367 |   |   |  
pagged-attn-fp8 | ShareGPT_V3_unfiltered_cleaned_split |   |   | 1000 | 6.61 | 2735.07 | 1311.8 | 10.53511706 | 10.5306548 | 10.53066177 | 21.50735294 | 21.66086179 | 21.6612257 |   |   |  



**BEFORE SUBMITTING, PLEASE READ https://docs.vllm.ai/en/latest/contributing/overview.html **