@tjtanaa 2025-01-24
This PR has been dropped for https://github.com/ROCm/vllm/pull/385 and https://github.com/ROCm/vllm/pull/372